+++ lib/cucumber/cli/options.rb
@@ -302,9 +302,8 @@ Specify SEED to reproduce the shuffling from a previous run.
           "formatter's docs to see whether to pass a file, dir or URL.",
           "\n",
           'When using a URL, the output of the formatter will be sent as the HTTP request body.',
-          'HTTP headers and request method can be set with http- prefixed query parameters,',
-          'for example ?http-content-type=application/json&http-method=PUT.',
-          'All http- prefixed query parameters will be removed from the sent query parameters.'
+          'HTTP headers and request method can be set with cURL like options.',
+          'Example: --out "http://example.com -X POST -H Content-Type:text/json"'
         ]
       end