+++ lib/cucumber/glue/proto_world.rb
@@ -107,9 +107,8 @@ module Cucumber
         attach(file, mime_type)
       end
 
-      def log(message)
-        raise Cucumber::LogTypeInvalid unless message.is_a?(String)
-        attach(message.dup, 'text/x.cucumber.log+plain')
+      def log(*messages)
+        messages.each { |message| attach(message.to_s.dup, 'text/x.cucumber.log+plain') }
       end
 
       def attach(file, media_type)