+++ spec/cucumber/glue/proto_world_spec.rb
@@ -58,6 +58,49 @@ OUTPUT
           end
         end
 
+        describe 'when logging an object' do
+          define_feature <<-FEATURE
+        Feature: Banana party
+
+          Scenario: Monkey eats banana
+            When an object is logged
+          FEATURE
+
+          define_steps do
+            When('an object is logged') do
+              log(a: 1, b: 2, c: 3)
+            end
+          end
+
+          it 'attached the styring version on the object' do
+            expect(@out.string).to include '{:a=>1, :b=>2, :c=>3}'
+          end
+        end
+
+        describe 'when logging multiple items on one call' do
+          define_feature <<-FEATURE
+        Feature: Banana party
+
+          Scenario: Monkey eats banana
+            When monkey eats banana
+          FEATURE
+
+          define_steps do
+            When('{word} {word} {word}') do |subject, verb, complement|
+              log "subject: #{subject}", "verb: #{verb}", "complement: #{complement}", subject: subject, verb: verb, complement: complement
+            end
+          end
+
+          it 'logs each parameter independently' do
+            expect(@out.string).to include [
+              '      subject: monkey',
+              '      verb: eats',
+              '      complement: banana',
+              '      {:subject=>"monkey", :verb=>"eats", :complement=>"banana"}'
+            ].join("\n")
+          end
+        end
+
         describe 'when modifying the printed variable after the call to log' do
           define_feature <<-FEATURE
         Feature: Banana party